py/requirements.txt
pathlib2

click
plumbum
gitpython
